WIP: restructure
