+++ crates/viewer/re_space_view_graph/src/ui/canvas.rs
@@ -3,14 +3,11 @@ use egui::{
     Stroke, Ui, Vec2,
 };
 use re_chunk::EntityPath;
-use re_types::{
-    components::{GraphNode, Radius},
-    datatypes::Float32,
-};
+use re_types::{components::Radius, datatypes::Float32};
 use re_viewer_context::SpaceViewHighlights;
 use std::hash::Hash;
 
-use crate::{graph::NodeInstanceImplicit, types::{EdgeInstance, NodeInstance}};
+use crate::{graph::NodeInstanceImplicit, visualizers::{EdgeInstance, NodeInstance}};
 
 use super::draw::{draw_edge, draw_entity, draw_explicit, draw_implicit};
